Improve engine part retention when fixing or loading
When fixing your car, engine parts will no longer drop. They will either be repaired, upgraded or added if missing.

When loading a preset, engine parts will only drop for modules that are changing. As before, the only parts that will drop are the ones that are of higher quality than allowed by the player's permissions.