Improve engine part retention when loading a preset
When loading a preset, existing engine parts of higher quality than allowed by the player's permissions will be redistributed among the engine modules in the loaded preset. Remaining parts will be added to the player inventory or dropped.

Despawning a car will also add engine parts to the player inventory or drop them. Same as with loading presets, this only applies to parts that are higher quality than allowed by the player's permissions.