Use Invoke instead of NextTick
Using Invoke is the correct approach to waiting until modules entities have been created since that is what the game uses to create them when the module items area added to the car inventory.

Using NextTick hasn't caused an issue in this plugin before, but it has been an issue in Car Spawn Settings. This may have something to do with whether the car is spawned in proximity to a player (or a subscribed network group), and it may be possible in the future of this plugin to spawn a car at specified coordinates, so making this change ahead of time.