Improved how spawn locations are determined
The plugin will now attempt to find terrain, construction or world objects to spawn on, using multiple raycasts at the projected car bounds, based on the car size. The car will also be rotated relative to where the raycasts hit.

Cars can no longer be spawned on players, inside buildings, or while the player is in the air or otherwise high above the surface that the car would spawn on.

As a side effect, some places are more difficult to spawn or fetch a car, such as when facing up a steep hill, but players are unlikely to attempt to spawn cars in these locations anyway, and they can usually change direction to spawn the car successfully, such as by facing downhill. Spawning on top of jagged rocks is also very difficult.